refactor: refactor scrolling behavior
